web_server_components/web_app/requirements.txt
Django>=3.1.3
Psycopg2>=2.8.6
Djangorestframework>=3.12.2
Pillow>=8.1.0
django_cors_headers>=3.7.0
django-filter>=2.4.0
django-crispy-forms>=1.11.1
django-csp>=3.7